backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
langchain==0.2.5
langchain-anthropic==0.1.15
langchain-pinecone==0.1.1
langchain-community==0.2.5
pinecone-client==3.2.2
pypdf==4.2.0
sentence-transformers==3.0.0
tiktoken==0.7.0
pydantic==2.7.4
pydantic-settings==2.3.1
python-dotenv==1.0.1
httpx==0.27.0
